requirements.txt
async-timeout==4.0.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
greenlet==1.1.2
httplib2==0.18.1
idna==2.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.0.11
MarkupSafe==2.1.1
packaging==21.3
pycurl==7.43.0.6
pyparsing==3.0.9
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
PyYAML==6.0
redis==4.3.1
reportbug==7.10.3+deb11u1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
